move parse_filename_like into cdf.utils
