+++ imap_processing/tests/cdf/test_utils.py
@@ -10,6 +10,7 @@ import xarray as xr
 from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
 from imap_processing.cdf.utils import (
     load_cdf,
+    parse_filename_like,
     write_cdf,
 )
 from imap_processing.spice.time import met_to_j2000ns
@@ -107,3 +108,55 @@ def test_parents_injection(test_dataset):
     parent_paths = [Path("test_parent1.cdf"), Path("/abc/test_parent2.cdf")]
     new_dataset = load_cdf(write_cdf(test_dataset, parent_files=parent_paths))
     assert new_dataset.attrs["Parents"] == [p.name for p in parent_paths]
+
+
+@pytest.mark.parametrize(
+    "test_str, compare_dict",
+    [
+        (
+            "imap_hi_l1b_45sensor-de",
+            {
+                "mission": "imap",
+                "instrument": "hi",
+                "data_level": "l1b",
+                "sensor": "45sensor",
+                "descriptor": "de",
+            },
+        ),
+        (
+            "imap_hi_l1a_hist_20250415_v001",
+            {
+                "mission": "imap",
+                "instrument": "hi",
+                "data_level": "l1a",
+                "descriptor": "hist",
+                "start_date": "20250415",
+                "version": "001",
+            },
+        ),
+        (
+            "imap_hi_l1c_90sensor-pset_20250415_v001.cdf",
+            {
+                "mission": "imap",
+                "instrument": "hi",
+                "data_level": "l1c",
+                "sensor": "90sensor",
+                "descriptor": "pset",
+                "start_date": "20250415",
+                "version": "001",
+                "extension": "cdf",
+            },
+        ),
+        ("foo_hi_l1c_90sensor-pset_20250415_v001.cdf", None),
+        ("imap_hi_l1c", None),
+    ],
+)
+def test_parse_filename_like(test_str, compare_dict):
+    """Test coverage for parse_filename_like function"""
+    if compare_dict:
+        match = parse_filename_like(test_str)
+        for key, value in compare_dict.items():
+            assert match[key] == value
+    else:
+        with pytest.raises(ValueError, match="Filename like string did not contain"):
+            _ = parse_filename_like(test_str)
